requirements.txt
pandas>=1.3.0
numpy>=1.21.0
beautifulsoup4>=4.10.0
requests>=2.26.0
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
python-dotenv>=0.19.0
pytest>=6.2.5
pytest-cov>=2.12.1
black>=21.12b0
isort>=5.10.1
flake8>=4.0.1
